Space Invaders beta
this is the start of this project I will add a starting GUI with a play and help buttons before it starts i will also change the boxes to pictures of the iconic space Invader alien. I will also make the spaceship look like the normal ship.  I will add more features soon