snake game (new) 1.0
Hey whats up, I have been working for a long time now if you look at the old snake game I made, it was totally different. I found that I was not using the right method so I scraped the code and re-did everything. The new code is using a grid system to move the snake and you can make hitboxes that way. I also changed how you die because in the old snake game you could not have an angle in the snake as it was going off the side of the map. Now it is the traditional way of hitting yourself. Thanks for the support of friends helping me make the code. I will be still making bug fixes and small changes but this will be the base code I will be making changes to. I hope you get to run the code and see the game I made, thanks for your time. 

<3 -panda